% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{Add}
\alias{Add}
\title{Add Op}
\usage{
Add(l_op, r_op, name = NULL)
}
\arguments{
\item{l_op}{Input node}

\item{r_op}{Input node}

\item{name}{Optional custom name for node, If not specified, a random name is generated}
}
\value{
Unique name of node
}
\description{
Returns x + y element-wise.
}
