% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{build_run_graph}
\alias{build_run_graph}
\title{Build and Run Graph}
\usage{
build_run_graph(feed, dtype = "double")
}
\arguments{
\item{feed}{Integer vector of size 3, feed for the network}

\item{dtype}{Datatype to be used, one of {"int32","double"}}
}
\value{
Output Value of Network
}
\description{
Function to build and run graph using C API
}
