% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_initializeSessionVariables}
\alias{c_initializeSessionVariables}
\title{Internal Initialize Session Variables}
\usage{
c_initializeSessionVariables()
}
\value{
Integer status
}
\description{
Use initializeSessionVariables() instead!
}
