% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_loadGraphFromFile}
\alias{c_loadGraphFromFile}
\title{Internal Load Graph from File}
\usage{
c_loadGraphFromFile(path)
}
\arguments{
\item{path}{Path to the graph}
}
\value{
Integer status
}
\description{
Use loadGraphFromFile() instead!
}
