% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_loadSavedModel}
\alias{c_loadSavedModel}
\title{Internal Function: Load Saved Model}
\usage{
c_loadSavedModel(path, tags)
}
\arguments{
\item{path}{Path to the saved model}

\item{tags}{Tags associated with the graph (from {"serve", "train, "gpu"})}
}
\value{
Integer Status
}
\description{
Use loadSavedModel() instead!
}
