% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_runSession}
\alias{c_runSession}
\title{Run Internal Session}
\usage{
c_runSession(op_names)
}
\arguments{
\item{op_names}{Node to be set as output of graph}
}
\value{
R List containing output tensor and dimensions
}
\description{
Runs the Current Interactive Session
}
