% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{check_mnist}
\alias{check_mnist}
\title{Load Saved Model}
\usage{
check_mnist(model_path, csv_path)
}
\arguments{
\item{model_path}{Path to saved model}

\item{csv_path}{Path to csv file with training data}
}
\value{
Output Value of Network
}
\description{
Load an MNIST graph and Train it. But first run mnist.py to create a saved model before running this function.
}
