% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deleteSessionVariables}
\alias{deleteSessionVariables}
\title{Close and Delete Session Variables}
\usage{
deleteSessionVariables()
}
\value{
Integer status
}
\description{
Closes session and frees all memory associated with it
}
\examples{
initializeSessionVariables()
deleteSessionVariables()

}
