% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{Equal}
\alias{Equal}
\title{Equal Op}
\usage{
Equal(l_op, r_op, name = NULL)
}
\arguments{
\item{l_op}{Input node}

\item{r_op}{Input node}

\item{name}{Optional custom name for node. If not specified, a random name is generated}
}
\value{
Unique name of node
}
\description{
Returns the truth value of (x == y) element-wise.
}
