% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{feedInput}
\alias{feedInput}
\title{Feed Input}
\usage{
feedInput(input_node, feed)
}
\arguments{
\item{input_node}{Node to which tensor must be fed to graph}

\item{feed}{Vector to be fed as input to the graph}
}
\value{
Integer status
}
\description{
Easy-to-use wrapper for setFeedInput
}
