% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{fetchOutput}
\alias{fetchOutput}
\title{Fetch Output}
\usage{
fetchOutput(dtype = "int32")
}
\arguments{
\item{dtype}{Datatype of output}
}
\value{
Output Array
}
\description{
Easy-to-use wrapper for getOutput
}
