% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{generateUniqueName}
\alias{generateUniqueName}
\title{Generate Unique Name}
\usage{
generateUniqueName(extra_length = 5, op_name = "")
}
\arguments{
\item{extra_length}{Number of characters to be randomly generated for identifier}

\item{op_name}{Name of op to be used as base of unique generated name}
}
\value{
Unique generated name for node
}
\description{
Generates a unique name for the node in the graph
}
