% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBinaryOp}
\alias{getBinaryOp}
\title{Binary Op}
\usage{
getBinaryOp(l_op, r_op, op_name, unique_name)
}
\arguments{
\item{l_op}{Input node}

\item{r_op}{Input node}

\item{op_name}{Type of operation for node}

\item{unique_name}{Unique name for the node}
}
\value{
Unique node name
}
\description{
Adds a binary operation to the graph
}
