% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getConstant}
\alias{getConstant}
\title{Constant}
\usage{
getConstant(val, dim, dtype, unique_name)
}
\arguments{
\item{val}{Tensor to be initialized as Constant}

\item{dim}{Vector indicating dimensions of val}

\item{dtype}{Datatype of input}

\item{unique_name}{Unique name for the node}
}
\value{
Unique node name
}
\description{
Adds a constant operation to the graph
}
