% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getOpProperties}
\alias{getOpProperties}
\title{Op Properties}
\usage{
getOpProperties(op_name)
}
\arguments{
\item{op_name}{Unique identifier of operation}
}
\value{
List with properties of op
}
\description{
Get properties of the operation
}
