% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getOutput}
\alias{getOutput}
\title{Fetch Output}
\usage{
getOutput(dtype)
}
\arguments{
\item{dtype}{Datatype to be fetched from output node}
}
\value{
List containing output vector and its dimensions
}
\description{
Fetches output of graph after running session
}
