% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getPlaceholder}
\alias{getPlaceholder}
\title{Placeholder}
\usage{
getPlaceholder(dtype, unique_name)
}
\arguments{
\item{dtype}{Datatype of input}

\item{unique_name}{Unique name for the node}
}
\value{
Unique node name
}
\description{
Adds a placeholder operation to the graph
}
