% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSourceOp}
\alias{getSourceOp}
\title{Source Op}
\usage{
getSourceOp(val, dim, dtype, op_name, unique_name)
}
\arguments{
\item{val}{Tensor to be initialized as Constant}

\item{dim}{Vector indicating dimensions of val}

\item{dtype}{Datatype of input}

\item{op_name}{Type of operation for node}

\item{unique_name}{Unique name for the node}
}
\value{
Unique node name
}
\description{
Adds a source operation to the graph
}
