% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getUnaryOp}
\alias{getUnaryOp}
\title{Unary Op}
\usage{
getUnaryOp(inp, op_name, unique_name)
}
\arguments{
\item{inp}{Input node}

\item{op_name}{Type of operation for node}

\item{unique_name}{Unique name for the node}
}
\value{
Unique node name
}
\description{
Adds a unary operation to the graph
}
