% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{import_run_graph}
\alias{import_run_graph}
\title{Import and Run Graph}
\usage{
import_run_graph(path, feed)
}
\arguments{
\item{path}{Path to the graph, "/tests/models/feed_forward_graph.pb"}

\item{feed}{Integer vector of size 3, feed for the network}
}
\value{
Output Value of Network
}
\description{
Function to import and run graph using C API
}
