% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{initializeSessionVariables}
\alias{initializeSessionVariables}
\title{Initialize Session Variables}
\usage{
initializeSessionVariables()
}
\value{
Integer status
}
\description{
Initializes all global variables for an interactive session
}
