% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{loadGraphFromFile}
\alias{loadGraphFromFile}
\title{Load Graph from File}
\usage{
loadGraphFromFile(path)
}
\arguments{
\item{path}{Path to the graph}
}
\value{
Integer status
}
\description{
Load graph from the file in the path specified
}
