% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{neural_net}
\alias{neural_net}
\title{Feed-Forward Network}
\usage{
neural_net(input, n_hidden)
}
\arguments{
\item{input}{Integer Vector for input to network, of any length.}

\item{n_hidden}{Integer for number of hidden neurons.}
}
\value{
int Output of Network
}
\description{
Simple feed forward network with X input neurons, 4 hidden neurons and 1 ouptut neuron.
}
