% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{Placeholder}
\alias{Placeholder}
\title{Placeholder Wrapper}
\usage{
Placeholder(dtype = "int32", name = "Placeholder")
}
\arguments{
\item{dtype}{Datatype of feed}

\item{name}{Optional custom name for node}
}
\value{
Unique name of node
}
\description{
Easy-to-use wrapper for getPlaceholder
}
