% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{Pow}
\alias{Pow}
\title{Pow Op}
\usage{
Pow(l_op, r_op, name = NULL)
}
\arguments{
\item{l_op}{Input node}

\item{r_op}{Input node}

\item{name}{Optional custom name for node. If not specified, a random name is generated}
}
\value{
Unique name of node
}
\description{
Computes the power of one value to another.
}
