% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{Relu}
\alias{Relu}
\title{Relu Op}
\usage{
Relu(inp, name = "Relu")
}
\arguments{
\item{inp}{Input node}

\item{name}{Optional custom name for node}
}
\value{
Unique name of node
}
\description{
Computes rectified linear: `max(features, 0)`.
}
