% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setFeedInput}
\alias{setFeedInput}
\title{Internal Feed Input Ops}
\usage{
setFeedInput(op_name, inp)
}
\arguments{
\item{op_name}{Op name (Node) to which tensor must be fed to graph}

\item{inp}{Vector to be fed to graph}
}
\value{
Integer status
}
\description{
Sets the input node of graph, and feeds a tensor to it
}
