% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setOutput}
\alias{setOutput}
\title{Set Output Ops}
\usage{
setOutput(op_name)
}
\arguments{
\item{op_name}{Op name (Node) to be set as output}
}
\value{
Integer status
}
\description{
Sets the output node of the graph
}
