% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{Sigmoid}
\alias{Sigmoid}
\title{Sigmoid Op}
\usage{
Sigmoid(inp, name = NULL)
}
\arguments{
\item{inp}{Input node}

\item{name}{Optional custom name for node. If not specified, a random name is generated}
}
\value{
Unique name of node
}
\description{
Computes sigmoid of `x` element-wise.
}
