% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{Softmax}
\alias{Softmax}
\title{Softmax Op}
\usage{
Softmax(inp, name = "Softmax")
}
\arguments{
\item{inp}{Input node}

\item{name}{Optional custom name for node}
}
\value{
Unique name of node
}
\description{
Computes softmax activations.
}
