% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{Add}
\alias{Add}
\title{Add Op}
\usage{
Add(l_op, r_op, name = "Add")
}
\arguments{
\item{l_op}{Input node}

\item{r_op}{Input node}

\item{name}{Optional custom name for node}
}
\value{
Unique name of node
}
\description{
Initializes an Add op in the graph. Returns x + y element-wise.
}
