% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{Constant}
\alias{Constant}
\title{Constant Wrapper}
\usage{
Constant(val, dim = c(length(val)), dtype = "int32", name = "Const")
}
\arguments{
\item{val}{Vector for value of Constant node}

\item{dim}{Vector indicating dimensions of val}

\item{dtype}{Datatype of val}

\item{name}{Optional custom name for node}
}
\value{
Unique name of node
}
\description{
Easy-to-use wrapper for getConstant
}
