% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{Constant}
\alias{Constant}
\title{Constant Wrapper}
\usage{
Constant(val, dtype = "int32", shape = c(length(val)), name = "Const")
}
\arguments{
\item{val}{Vector for value of Constant node}

\item{dtype}{Datatype of val}

\item{shape}{Vector indicating dimensions of val}

\item{name}{Optional custom name for node}
}
\value{
Unique name of node
}
\description{
Easy-to-use wrapper for getConstant. Returns a constant tensor.
}
