% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{Cos}
\alias{Cos}
\title{Cos Op}
\usage{
Cos(inp, name = NULL)
}
\arguments{
\item{inp}{Input node}

\item{name}{Optional custom name for node. If not specified, a random name is generated}
}
\value{
Unique name of node
}
\description{
Computes cos of x element-wise.
}
