% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{feedInput}
\alias{feedInput}
\title{Feed Input}
\usage{
feedInput(input_node, feed, dim = c(length(feed)), dtype = "int32")
}
\arguments{
\item{input_node}{Node to be set as input to graph}

\item{feed}{Vector to be fed as input to the graph}

\item{dim}{Vector indicating dimensions of feed}

\item{dtype}{Datatype of feed}
}
\value{
Integer status
}
\description{
Easy-to-use wrapper for setFeedInput
}
