% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loadGraphFromFile}
\alias{loadGraphFromFile}
\title{Load Graph from File}
\usage{
loadGraphFromFile(path)
}
\arguments{
\item{path}{Path to the graph}
}
\value{
Integer status
}
\description{
Assigns graph variable with one indicated by path
}
\examples{
loadGraphFromFile("/tests/models/feed_forward_graph.pb")

}
