% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{MatMul}
\alias{MatMul}
\title{MatMul Op}
\usage{
MatMul(l_op, r_op, name = "MatMul")
}
\arguments{
\item{l_op}{Input node}

\item{r_op}{Input node}

\item{name}{Optional custom name for node}
}
\value{
Unique name of node
}
\description{
Initializes a MatMul op in the graph. Multiply the matrix x by the matrix y.
}
