% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{Neg}
\alias{Neg}
\title{Neg Op}
\usage{
Neg(inp, name = "Neg")
}
\arguments{
\item{inp}{Input node}

\item{name}{Optional custom name for node}
}
\value{
Unique name of node
}
\description{
Computes numerical negative value element-wise.
}
