% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{Placeholder}
\alias{Placeholder}
\title{Placeholder Wrapper}
\usage{
Placeholder(dtype, shape = NULL, name = NULL)
}
\arguments{
\item{dtype}{Shape of Tensor}

\item{shape}{Datatype of Tensor}

\item{name}{Optional custom name for node. If not specified, a random name is generated}
}
\value{
Unique name of node
}
\description{
Sets a placeholder op for a value that will be fed into the computation.
}
