% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{Placeholder}
\alias{Placeholder}
\title{Placeholder Wrapper}
\usage{
Placeholder(dtype, shape = NULL, name = "Placeholder")
}
\arguments{
\item{dtype}{Shape of Tensor}

\item{shape}{Datatype of Tensor}

\item{name}{Optional custom name for node}
}
\value{
Unique name of node
}
\description{
Easy-to-use wrapper for getPlaceholder. Sets a placeholder op for a value that will be fed into the computation.
}
