% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{runSession}
\alias{runSession}
\title{Run Interactive Session}
\usage{
runSession(op_names)
}
\arguments{
\item{op_names}{Node to be set as output of graph}
}
\value{
Multidimensional output matrix
}
\description{
Runs the current Tensorflow session
}
