% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setFeedInput}
\alias{setFeedInput}
\title{Feed Input Ops}
\usage{
setFeedInput(op_name, inp, dim, dtype)
}
\arguments{
\item{op_name}{Op name (Node) to be set as input to graph}

\item{inp}{Vector to be fed to graph}

\item{dim}{Vector indicating dimensions of inp}

\item{dtype}{Datatype of inp}
}
\value{
Integer status
}
\description{
Sets the input node of graph, and feeds a tensor to it
}
