% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_api.R
\name{Tanh}
\alias{Tanh}
\title{Tanh Op}
\usage{
Tanh(inp, name = "Tanh")
}
\arguments{
\item{inp}{Input node}

\item{name}{Optional custom name for node}
}
\value{
Unique name of node
}
\description{
Computes hyperbolic tangent of `x` element-wise.
}
